backend/requirements.txt
fastapi
uvicorn
sqlmodel
alembic
pydantic_settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart
email-validator
psycopg2-binary
reportlab
Pillow
google-cloud-storage
jinja2
weasyprint